+++ packages/web3-eth-accounts/CHANGELOG.md
@@ -138,3 +138,7 @@ Documentation:
 
 
 ## [Unreleased]
+
+### Added
+
+-   Added public function `privateKeyToPublicKey`